Python 3 compatibility (using six); has_strings
The `has_strings` function searches a haystack string for several
needles; if anything is found, an information is printed (default `func`
argument is the `print` function), and `True` is returned.

This is aimed for debugging, of course, e.g. for use as a breakpoint
criterion.

There is a `make_debugfile_writer` function as well; but it is not yet
sufficiently generalized, thus it is not "published" in the `__all__`
list.