Restored compatibility of the .lock module
In the py2+3 branch, we maintain Python 2.7 compatibility,
so some changes were applied to achieve this.

The check_lockfile_options helper function doesn't enforce the use of
kwargs here, because there is no way to have Python 2.7 understand the
'*' formal argument.

We used to have f-strings here as well (PEP 498), which we replaced by
classic Python %(...)s string interpolation.
Thus, you may use this branch with elderly Python 3 versions (3.3+?) as
well.
It should work for current Python as well (we saw it work with Python
3.12.3), but you'll likely prefer the py3-only branch if possible.